# cmd/api/docs/swagger.yaml
definitions:
...
  helperStruct.OTPData:
    properties:
      phoneNumber:
        type: string
    required:
    - phoneNumber
    type: object
...
  helperStruct.VerifyOtp:
    properties:
      code:
        type: string
      user:
        $ref: '#/definitions/helperStruct.OTPData'
    required:
    - code
    - user
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /addtofav/{productId}:
    post:
      consumes:
      - application/json
      description: User can add product item to favourites
      operationId: add-to-favourites
      parameters:
      - description: ID of the product item to be added to wishlist
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can add product item to favourites
      tags:
      - Favourites
...
  /listallorder:
    get:
      consumes:
      - application/json
      description: Endpoint for getting all orders associated with a user
      operationId: view-all-orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Retrieves all orders of currently logged in user
      tags:
      - Order
...
  /orders/cancel/{orderId}:
    patch:
      consumes:
      - application/json
      description: Endpoint for cancelling an order associated with a user
      operationId: cancel-order
      parameters:
      - description: ID of the order to be cancelled
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Cancels a specific order for the currently logged in user
      tags:
      - Order
...
  /removefromfav/{productId}:
    delete:
      consumes:
      - application/json
      description: User can remove product item from favourites
      operationId: remove-from-favourites
      parameters:
      - description: ID of the product item to be added to wishlist
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can remove product item from favourites
      tags:
      - Favourites
  /return/{orderId}:
    patch:
      consumes:
      - application/json
      description: Endpoint for Returning an order associated with a user
      operationId: return-order
      parameters:
      - description: ID of the order to be cancelled
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Return a specific order for the currently logged in user
      tags:
      - Order
  /sendotp:
    post:
      consumes:
      - application/json
      description: Send OTP to use's mobile
      operationId: send-otp
      parameters:
      - description: User mobile number
        in: body
        name: user_mobile
        required: true
        schema:
          $ref: '#/definitions/helperStruct.OTPData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Send OTP to user's mobile
      tags:
      - Otp
...
  /verifyotp:
    post:
      consumes:
      - application/json
      description: Validate the  OTP sent to use's mobile
      operationId: validate-otp
      parameters:
      - description: OTP sent to user's mobile number
        in: body
        name: otp
        required: true
        schema:
          $ref: '#/definitions/helperStruct.VerifyOtp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Validate the OTP to user's mobile
      tags:
      - Otp
  /viewfav/:
    get:
      consumes:
      - application/json
      description: User view product items in favourites
      operationId: view-favourites
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can view items in favourites
      tags:
      - Favourites
  /vieworder/{orderId}:
    get:
      consumes:
      - application/json
      description: This function handles requests for retrieving the details of a
        specific order identified by its order ID.
      operationId: view-order-by-id
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched order details
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to fetch order details
          schema:
            $ref: '#/definitions/response.Response'
      summary: Retrieves order details for a given order ID, if authorized.
      tags:
      - Order
...
swagger: "2.0"